core: use NMPtrArray in nm_netns_ip_route_ecmp_commit() to track list of pointers
In this scenario, NMPtrArray serves the same purpose as GPtrArray,
albeit better. For the first 8 pointers it requires no heap allocation,
before it reallocates a buffer on the heap. Unlike GPtrArray which
requires an extra allocation for the GPtrArray structure and already
requires to allocate a buffer for a single pointer.