ovs: rename internal code to make independent of "external-ids"
We will add support for "other_config". This is in many aspects similar
to "external-ids". So first do a renaming, so that the code can be
sensibly reused. This is a separate patch, so that the followup commit
has less noise in the diff.

This function *only* renames (and reformats). No other changes.