all: don't add proxy setting during normalization
It doesn't seem useful to always add the proxy setting to each
connection profile that supports IP configuration. This normalization
requires additional code and especially, it creates a new empty
section in every new keyfile:

 # nmcli connection add type dummy ifname dummy1
 # cat /etc/NetworkManager/system-connections/dummy-dummy1.nmconnection
 [connection]
 id=dummy-dummy1
 uuid=8dd2bb73-b338-428f-ad6d-082c37a58566
 type=dummy
 interface-name=dummy1

 [dummy]

 [ipv4]
 method=disabled

 [ipv6]
 addr-gen-mode=default
 method=disabled

 [proxy]

Since an empty (with method=none) proxy setting is the same as no
proxy setting, avoid the normalization.