core: remove NM_PROXY_CONFIG_METHOD_UNKNOWN
I don't see a reason to have NM_PROXY_CONFIG_METHOD_UNKNOWN, as it is
treated exactly the same as NM_PROXY_CONFIG_METHOD_NONE, i.e. they
both send a "direct" method to pacrunner. The only difference is that
UNKNOWN indicates the absence of the proxy setting in the profile, but
that is the same as having the setting with method=none.