fixup! policy: track autoconnect retries per Device x Connection
The function is-blocked performs a series of checks, and when any of
them is true, it returns is-blocked TRUE. It's a list of "if(check)return
TRUE". Follow that pattern consistently throughout the function, and end
with a return FALSE.

Also, drop the blank line between checking "flags" for condition1 and
condition2.

Beside, make the code as it was previously.