vapi: Add several namespace changes and more specific types for properties
Add overrides for properties constants of new objects to allow to match them with their
object type as it was the case for former objects.

Specify the types of properties with element-types.

Namespace Attributes so that they are logically sorted.

Remove several skips to allow to use the entirety of the visible methods.